ceph: permit builds without a VERSION
Prior to this change, we always had to pass a VERSION to the ceph job.
The job would create a tag based on this VERSION number, push the tag to
a private repo, and build it.

With this change, when the user does not specify a VERSION, Jenkins will
simply build the BRANCH directly, without tagging or using the private
repo.

When building without a tag, the ceph-setup job does not generate a
.diff.gz for some reason. It's probably a bug that we generate this
file during any build, at all :( For now, just work around this issue by
updating the "mv" command so that we only try to move .diff.gz if the
file exists.